Update demo ch11_1
The demo is updated to better match the R-demos.
Animations for the samplings were added. A lot of comments
and markup sections were added to more thoroughly explain
what is done in the demo. Also the errors in labeling and
legends were corrected. Also contradictions between markup
sections and code sections were corrected. When doing this demo
we spotted similar errors in the R-demos. For example
the markup sections sometimes talk about 50 burnin samples
when the code actually uses 500 burnin samples etc.